Fix list of allowed verbs for failed preflight
During preflight (OPTIONS verb) no Authorization headers can be
provided. Some apps requirese these headers to determine the identity,
when solving a path and associating it with a model (eg: imagine that
the model is a User whose identity is in that header). Without that
information a sensible action for the path function is to raise
HTTPUnauthorized.

The problem is that more.cors treats that exception (and any exception)
in the same way during preflight: a 404 error is returned and no cors
headers. This prevents the browser to continue with the operation, which
could be otherwise legal.

This patch causes more.cors to return valid cors headers even if
the path function raised HTTPUnauthorized, instead of no cors headers.